improvement(refactor-structure): Refactor project's structure so by creating a package per screen including its corresponding components package. Delete unused ChipCategoryItem
