+++ app/src/main/java/com/enriqueajin/newsapp/data/network/model/NewsItem.kt
@@ -1,4 +1,4 @@
-package com.enriqueajin.newsapp.ui.model
+package com.enriqueajin.newsapp.data.network.model
 
 import kotlinx.serialization.Serializable
