+++ app/src/main/java/com/enriqueajin/newsapp/ui/home/HomeScreen.kt
@@ -8,10 +8,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
-import com.enriqueajin.newsapp.ui.NewsViewModel
-import com.enriqueajin.newsapp.ui.home.components.bottom_navigation.BottomNav
-import com.enriqueajin.newsapp.ui.home.components.tab_row.TopTabRow
-import com.enriqueajin.newsapp.ui.model.NewsItem
+import com.enriqueajin.newsapp.ui.home.tabs.news.NewsViewModel
+import com.enriqueajin.newsapp.ui.home.components.BottomNav
+import com.enriqueajin.newsapp.ui.home.components.TopTabRow
+import com.enriqueajin.newsapp.data.network.model.NewsItem
 
 @Composable
 fun Home(newsViewModel: NewsViewModel, onSeeAllClicked: (List<NewsItem>) -> Unit) {
