+++ app/src/main/java/com/enriqueajin/newsapp/ui/home/components/TopTabRow.kt
@@ -1,4 +1,4 @@
-package com.enriqueajin.newsapp.ui.home.components.tab_row
+package com.enriqueajin.newsapp.ui.home.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -20,11 +20,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.enriqueajin.newsapp.ui.NewsViewModel
+import com.enriqueajin.newsapp.ui.home.tabs.news.NewsViewModel
 import com.enriqueajin.newsapp.ui.home.tabs.Events
-import com.enriqueajin.newsapp.ui.home.tabs.News
+import com.enriqueajin.newsapp.ui.home.tabs.news.News
 import com.enriqueajin.newsapp.ui.home.tabs.Weather
-import com.enriqueajin.newsapp.ui.model.NewsItem
+import com.enriqueajin.newsapp.data.network.model.NewsItem
 import kotlinx.coroutines.launch
 
 @Composable
