--- app/src/main/java/com/enriqueajin/newsapp/ui/home/tabs/LocalUiState.kt
@@ -1,6 +1,0 @@
-package com.enriqueajin.newsapp.ui.home.tabs
-
-data class LocalUiState(
-    val categorySelected: String = "All",
-    val keyword: String = "Tesla"
-)
