+++ app/src/main/java/com/enriqueajin/newsapp/ui/home/tabs/news/NewsLocalUiState.kt
@@ -1,0 +1,6 @@
+package com.enriqueajin.newsapp.ui.home.tabs.news
+
+data class NewsLocalUiState(
+    val categorySelected: String = "All",
+    val keyword: String = "Tesla"
+)
