+++ app/src/main/java/com/enriqueajin/newsapp/ui/home/tabs/news/components/all_news/LatestNewsItem.kt
@@ -1,4 +1,4 @@
-package com.enriqueajin.newsapp.ui.home.components.latest_news_carousel
+package com.enriqueajin.newsapp.ui.home.tabs.news.components.all_news
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.enriqueajin.newsapp.ui.model.NewsItem
+import com.enriqueajin.newsapp.data.network.model.NewsItem
 import com.enriqueajin.newsapp.ui.theme.Purple80
 import com.enriqueajin.newsapp.util.DummyDataProvider.getLatestNewsItems
