+++ app/src/main/java/com/enriqueajin/newsapp/ui/home/tabs/news/components/keyword_news/NewsByCategory.kt
@@ -1,0 +1,11 @@
+package com.enriqueajin.newsapp.ui.home.tabs.news.components.keyword_news
+
+import androidx.compose.foundation.lazy.LazyListScope
+import androidx.compose.foundation.lazy.items
+import com.enriqueajin.newsapp.data.network.model.NewsItem
+
+fun LazyListScope.NewsByCategory(news: List<NewsItem>) {
+    items(news) { item ->
+        NewsListItem(item = item)
+    }
+}
