--- app/src/main/java/com/enriqueajin/newsapp/ui/model/ChipCategoryItem.kt
@@ -1,7 +1,0 @@
-package com.enriqueajin.newsapp.ui.model
-
-data class ChipCategoryItem(
-    val label: String,
-    var selected: Boolean = false,
-    val onClick: () -> Unit
-)
